refactor : 계산 관련 함수 Calcaultor 클래스로 분리
